Testes

+++ source/Estrutura_de_dados/Estrutura_de_dados.hpp
@@ -13,10 +13,10 @@ class EstruturaDeDados{
 
     virtual void setSize(int newSizeVector)=0;
     virtual bool addAresta(int valor1, int valor2)=0;
-    virtual bool addAresta(int valor1, int valor2, float weight)=0;
+    virtual bool addAresta(int valor1, int valor2, double weight)=0;
     virtual void show(bool weight)=0;
     virtual int vizinhoDeVertice(int vertice, int posicaoVizinho)=0;
-    virtual pair <int,float> vizinhoDeVertice(int vertice, int posicaoVizinho, bool weight)=0;
+    virtual pair <int,double> vizinhoDeVertice(int vertice, int posicaoVizinho, bool weight)=0;
     virtual int sizeVertice(int vertice)=0;
     //virtual void addVertice();
     //virtual bool addAresta(int valor1, int valor2);
